+++ Source/Classes/PINRemoteImageDownloadTask.m
@@ -80,9 +80,10 @@
                 progressImageBlock([PINRemoteImageManagerResult imageResultWithImage:image
                                                            alternativeRepresentation:nil
                                                                        requestLength:CACurrentMediaTime() - requestTime
-                                                                               error:nil
                                                                           resultType:PINRemoteImageResultTypeProgress
                                                                                 UUID:UUID
+                                                                            response:nil
+                                                                               error:nil
                                                                 renderedImageQuality:renderedImageQuality]);
            });
         }
@@ -156,6 +157,7 @@
                                                         error:(nullable NSError *)error
                                                    resultType:(PINRemoteImageResultType)resultType
                                                          UUID:(nullable NSUUID *)UUID
+                                                     response:(nonnull NSURLResponse *)response
 {
     __block NSUInteger bytesSavedByResuming;
     [self.lock lockWithBlock:^{
@@ -164,9 +166,10 @@
     return [PINRemoteImageManagerResult imageResultWithImage:image
                                    alternativeRepresentation:alternativeRepresentation
                                                requestLength:requestLength
-                                                       error:error
                                                   resultType:resultType
                                                         UUID:UUID
+                                                    response:response
+                                                       error:error
                                         bytesSavedByResuming:bytesSavedByResuming];
 }
 
@@ -337,7 +340,7 @@
                         }
                     }
                     
-                    completionHandler(data, error);
+                    completionHandler(data, response, error);
                 }
             }];
         }]];