+++ Source/Classes/PINRemoteImageManagerResult.m
@@ -10,28 +10,31 @@
 
 @implementation PINRemoteImageManagerResult
 
-+ (instancetype)imageResultWithImage:(PINImage *)image
-           alternativeRepresentation:(id)alternativeRepresentation
-                       requestLength:(NSTimeInterval)requestLength
-                               error:(NSError *)error
-                          resultType:(PINRemoteImageResultType)resultType
-                                UUID:(NSUUID *)uuid
++ (nonnull instancetype)imageResultWithImage:(nullable PINImage *)image
+                   alternativeRepresentation:(nullable id)alternativeRepresentation
+                               requestLength:(NSTimeInterval)requestLength
+                                  resultType:(PINRemoteImageResultType)resultType
+                                        UUID:(nullable NSUUID *)uuid
+                                    response:(nullable NSURLResponse *)response
+                                       error:(nullable NSError *)error
 {
     return [self imageResultWithImage:image
             alternativeRepresentation:alternativeRepresentation
                         requestLength:requestLength
-                                error:error
                            resultType:resultType
                                  UUID:uuid
+                             response:response
+                                error:error
                  renderedImageQuality:1.0];
 }
 
 + (nonnull instancetype)imageResultWithImage:(nullable PINImage *)image
                    alternativeRepresentation:(nullable id)alternativeRepresentation
                                requestLength:(NSTimeInterval)requestLength
-                                       error:(nullable NSError *)error
                                   resultType:(PINRemoteImageResultType)resultType
                                         UUID:(nullable NSUUID *)uuid
+                                    response:(nullable NSURLResponse *)response
+                                       error:(nullable NSError *)error
                         bytesSavedByResuming:(NSUInteger)bytesSavedByResuming
 {
     return [[self alloc] initWithImage:image
@@ -40,17 +43,19 @@
                                  error:error
                             resultType:resultType
                                   UUID:uuid
+                              response:response
                   renderedImageQuality:1.0
                   bytesSavedByResuming:bytesSavedByResuming];
 }
 
-+ (instancetype)imageResultWithImage:(PINImage *)image
-           alternativeRepresentation:(id)alternativeRepresentation
-                       requestLength:(NSTimeInterval)requestLength
-                               error:(NSError *)error
-                          resultType:(PINRemoteImageResultType)resultType
-                                UUID:(NSUUID *)uuid
-                renderedImageQuality:(CGFloat)renderedImageQuality
++ (nonnull instancetype)imageResultWithImage:(nullable PINImage *)image
+                   alternativeRepresentation:(nullable id)alternativeRepresentation
+                               requestLength:(NSTimeInterval)requestLength
+                                  resultType:(PINRemoteImageResultType)resultType
+                                        UUID:(nullable NSUUID *)uuid
+                                    response:(nullable NSURLResponse *)response
+                                       error:(nullable NSError *)error
+                        renderedImageQuality:(CGFloat)renderedImageQuality
 {
     return [[self alloc] initWithImage:image
              alternativeRepresentation:alternativeRepresentation
@@ -58,6 +63,7 @@
                                  error:error
                             resultType:resultType
                                   UUID:uuid
+                              response:response
                   renderedImageQuality:renderedImageQuality
                   bytesSavedByResuming:0];
 }
@@ -68,6 +74,7 @@
                         error:(NSError *)error
                    resultType:(PINRemoteImageResultType)resultType
                          UUID:(NSUUID *)uuid
+                     response:(NSURLResponse *)response
          renderedImageQuality:(CGFloat)renderedImageQuality
          bytesSavedByResuming:(NSUInteger)bytesSavedByResuming;
 {
@@ -78,6 +85,7 @@
         _error = error;
         _resultType = resultType;
         _UUID = uuid;
+        _response = response;
         _renderedImageQuality = renderedImageQuality;
         _bytesSavedByResuming = bytesSavedByResuming;
     }
@@ -99,7 +107,11 @@
     description = [description stringByAppendingString:@"\n"];
     description = [description stringByAppendingString:[NSString stringWithFormat:@"UUID: %@", self.UUID]];
     description = [description stringByAppendingString:@"\n"];
+    description = [description stringByAppendingString:[NSString stringWithFormat:@"response: %@", self.response]];
+    description = [description stringByAppendingString:@"\n"];
     description = [description stringByAppendingString:[NSString stringWithFormat:@"renderedImageQuality: %f", self.renderedImageQuality]];
+    description = [description stringByAppendingString:@"\n"];
+    description = [description stringByAppendingString:[NSString stringWithFormat:@"bytesSavedByResuming: %lu", (unsigned long)self.bytesSavedByResuming]];
     return description;
 }